+++ src/ConfigProvider.php
@@ -9,14 +9,14 @@ namespace XDevApi;
  */
 final class ConfigProvider
 {
-    public function __invoke() : array
+    public function __invoke(): array
     {
         return [
             'dependencies' => $this->getDependencies()
         ];
     }
 
-    public function getDependencies() : array
+    public function getDependencies(): array
     {
         return [
             'factories'  => [