+++ test/Entity/DocumentEntityTest.php
@@ -1,4 +1,6 @@
-<?php /** @noinspection ALL */
+<?php
+
+/** @noinspection ALL */
 
 declare(strict_types=1);
 
@@ -71,4 +73,3 @@ class DocumentEntityTest extends TestCase
         $this->assertSame('{"test_field":"test"}', json_encode($entity));
     }
 }
-